rf: move `get_linkml_err_counts()` and supporting def
From `cli/tools.py` to `tools.py`
