+++ src/dandisets_linkml_status_tools/tools.py
@@ -3,8 +3,9 @@ from collections import Counter
 from collections.abc import Iterable
 from copy import deepcopy
 from functools import partial
+from itertools import chain
 from pathlib import Path
-from typing import Any
+from typing import Any, NamedTuple
 
 from dandi.dandiapi import RemoteDandiset
 from dandischema.models import PublishedDandiset, Dandiset
@@ -21,7 +22,7 @@ from .cli.tools import DANDI_MODULE_NAMES
 from .models import (
     ValidationReport,
     DandisetLinkmlTranslationReport,
-    PydanticValidationErrsType,
+    PydanticValidationErrsType, JsonschemaValidationErrorType, LinkmlValidationErrsType,
 )
 
 logger = logging.getLogger(__name__)
@@ -235,3 +236,72 @@ def get_pydantic_err_counts(errs: PydanticValidationErrsType) -> Counter[str]:
     :return: The `Counter` object
     """
     return Counter(isorted(e["type"] for e in errs))
+
+
+class _JsonschemaValidationErrorCounts(NamedTuple):
+    """
+    A record of the counts of individual types of JSON schema validation error
+    """
+
+    types: list[JsonschemaValidationErrorType]
+    """
+    The unique types of JSON schema validation errors
+    """
+
+    counts: list[int]
+    """
+    The corresponding counts, by index, of the types of JSON schema validation errors
+    """
+
+
+def get_linkml_err_counts(
+    errs: LinkmlValidationErrsType,
+) -> list[tuple[JsonschemaValidationErrorType, int]]:
+    """
+    Counts given LinkML validation errors by type
+
+    :param errs: A list of LinkML validation errors to be counted
+    :return: A list of tuples where each tuple contains a
+        `JsonschemaValidationErrorType` object and the count of the errors of the type
+        represented by that object
+    """
+
+    def count_err(e_: ValidationResult) -> None:
+        validator = e_.source.validator
+        err_type = JsonschemaValidationErrorType(validator, e_.source.validator_value)
+
+        if validator in counter:
+            for i, t in enumerate(counter[validator].types):
+                if t == err_type:
+                    counter[validator].counts[i] += 1
+                    break
+            else:
+                counter[validator].types.append(err_type)
+                counter[validator].counts.append(1)
+        else:
+            counter[validator] = _JsonschemaValidationErrorCounts(
+                types=[err_type], counts=[1]
+            )
+
+    def compile_counts() -> list[tuple[JsonschemaValidationErrorType, int]]:
+        def sorting_key(
+            c: tuple[JsonschemaValidationErrorType, int],
+        ) -> tuple[str, int]:
+            return c[0].validator, -c[1]
+
+        return sorted(
+            chain.from_iterable(zip(t, c, strict=False) for t, c in counter.values()),
+            key=sorting_key,
+        )
+
+    # A dictionary that keeps the counts of individual types of JSON schema validation
+    # errors. The keys of the dictionary are the `validator` of
+    # the `JsonschemaValidationErrorType` objects, and the values are
+    # the `_JsonschemaValidationErrorCounts` that tallies the errors represented by
+    # `JsonschemaValidationErrorType` objects with the same `validator` value.
+    counter: dict[str, _JsonschemaValidationErrorCounts] = {}
+
+    for e in errs:
+        count_err(e)
+
+    return compile_counts()
